Move bootloader functions to bootloader.c / bootloader.h
Bootloader specific functions moved from main to individual
files:
bootloader.c
bootloader.h
